adding a KF for smoothing
