+++ pose_publisher/launch/pose_publisher.launch
@@ -1,5 +1,5 @@
 <launch>
-    <param name="config_file" value="/home/subodh/ardupilot_ws/src/pose_publisher/data/cam_sv_bl.yaml" /> 
+    <param name="config_file" value="/home/subodh/catkin_ws/src/pose_publisher/data/cam_sv_bl.yaml" /> 
     <node name="pose_publisher" pkg="pose_publisher" type="pose_publisher_node" output="screen">
     </node>
 </launch>
